In B-tree commit, moved pending attrs before merge attempts
I'm still not sure this is the right place for this, but it does
simplify pending attr calculations during merge and deduplicates
two instances of writing pending attrs, at the cost of needing to
track an additional rbyd weight during merge.

Going to roll with this for now, the B-tree merge code needs to be
cleaned up anyways, maybe it's possible to simplify the state we need to
track.

Another side-effect is this makes our B-trees slightly less aggressive
at merging. I have no idea if this is a good or bad thing.