Flipped layout of fcrcs to match B-tree branches
This may seem more complicated to decode, we can't assume crcs start
at the beginning of the data, but this layout of putting the crcs at the
end has the benefit of allowing the size of the crc to be unknown in
certain cases.

The is a bit of optimistic future proofing for the case where we may
support different crc widths.