Implemented and adopted rbyd compaction estimates
Still needs work, but at least adopted optionally in the btree.

Ignoring the mdirs for now, which is a bit ironic, because the mdir
compaction is really what this feature is for. But this at least proves
the concept.

---

Unlike btrees, mdirs simply cannot perform the attempt-then-delete-half
strategy current performed by the btrees during compaction with a single
pcache. This is because the moment we finish the commit with the delete,
it becomes visible to the filesystem. We can't abort the commit temporarily
to deal with the other half of the split, because our pcache is in use.

So, instead, the idea is to estimate the compacted rbyd size before
compacting, using conservative (but tight!) estimates for various leb128
encoded parts of the metadata.

And if we adopt this strategy for mdirs, we should probably adopt it in
the btrees for better code sharing.

A couple benefits:

- Major reduction in progs during split, since we don't write out tags
  just to delete them.

- btree merge can actually consider both siblings now.

- Not needing to weave the split/merge logic around compact offers a
  better route for code deduplication.

- mdir compact will actually work, that's generally a good thing.

And a couple downsides:

- This estimate is complex, meaning more code-cost and a bigger surface
  area for bugs.

- This results in a minor performance hit for the common compact case,
  since we need to read the rbyd being compacted twice instead of once.