Simplified btree merge estimate at the risk of failed merges
This trades off simpler estimation (and more flexibility for
experimenting with better compaction strategies) for the possibility of
failed merges that require cleanup.

Note, we still estimate if our merge fits post-compaction, this just
isn't reliable for knowing if the merge fits appended to the current
compaction, with in-flight attrs for both the in-flight commit and merge name.

Fortunately, since our estimate is conservative, we shouldn't see any
split<->merge oscillation.

---

It's also worth noting since the merge name isn't accounted for in our
sibling, it wasn't clear if our estimation was correct in the first
place.

This change avoids any issues that may cause.

One downside of using our compaction estimate as a heuristic to avoid
failed merges: The merge abort code path may be difficult to cover in tests.

We should make sure merge aborts don't go untested.