Tweaked lfsr_mdir_compact_ to make source rbyd optional
This allows lfsr_mdir_compact_ to also cover the mroot extension commit.

The mroot located at 0x{0,1} is a bit unique in that it can never
relocated. Instead we "extend" the mroot chain by an additional mroot
that can relocate.

One nice thing is we can implement this by letting lfsr_mdir_commit_
perform a normal relocation, and then rewrite the 0x{0,1} mroot with
a pointer to the "relocated" mroot.

Though we have to take extra care to make sure this write doesn't
recursively trigger an additional relocate, which would never terminate.

Fortunately, this situation only happens when we are compacting the
0x{0,1} mroot. Which simplifies things a bit.